refactor(llcppsigfetch):reuse llgo/preprocessor
