+++ _xtool/internal/parser/_testdata/hfile/compat.h
