+++ _xtool/internal/parser/_testdata/hfile/main.h
