+++ _xtool/llcppsigfetch/internal/parse/parse.go
@@ -10,6 +10,7 @@ import (
 	"github.com/goplus/llcppg/_xtool/internal/config"
 	"github.com/goplus/llcppg/_xtool/internal/parser"
 	llcppg "github.com/goplus/llcppg/config"
+	"github.com/goplus/llgo/xtool/clang/preprocessor"
 )
 
 type dbgFlags = int
@@ -79,16 +80,18 @@ func Do(conf *Config) error {
 
 	// prepare clang flags to preprocess the combined file
 	clangFlags := strings.Fields(conf.Conf.CFlags)
+	if !isCpp {
+		clangFlags = append(clangFlags, "-x", "c")
+	}
 	clangFlags = append(clangFlags, "-C")  // keep comment
 	clangFlags = append(clangFlags, "-dD") // keep macro
 	clangFlags = append(clangFlags, "-fparse-all-comments")
 
-	err = clangtool.Preprocess(&clangtool.PreprocessConfig{
-		File:    conf.CombinedFile,
-		IsCpp:   isCpp,
-		Args:    clangFlags,
-		OutFile: conf.PreprocessedFile,
-	})
+	ppconf := &preprocessor.Config{
+		Compiler: "clang",
+		Flags:    clangFlags,
+	}
+	err = preprocessor.Do(conf.CombinedFile, conf.PreprocessedFile, ppconf)
 	if err != nil {
 		return err
 	}
